Update: routes and variables renaming and path changes
