+++ src/app/pages/CChainPages/Blocks/index.tsx
@@ -14,7 +14,7 @@ import BackButton from 'app/components/BackButton';
 import { useInfiniteQuery } from 'react-query';
 
 import Block from './Block';
-import CutomTable from 'app/components/Table/CustomTable';
+import CutomTable from 'app/components/Table/TableView';
 import useWidth from 'app/hooks/useWidth';
 import { getBlocksPage } from 'api';
