+++ src/utils/magellan-api-utils.ts
@@ -8,12 +8,12 @@ export const cBlocksApi = baseEndpoint + '/cblocks';
 
 export const cBlocksDetailsApi = baseEndpoint + '/ctxdata';
 
-export const search = baseEndpoint + '/search';
+export const searchApi = baseEndpoint + '/search';
 
-export const addresses = baseEndpoint + '/addresses';
+export const addressesApi = baseEndpoint + '/addresses';
 
-export const assets = baseEndpoint + '/assets';
+export const assetsApi = baseEndpoint + '/assets';
 
-export const transactionAggregates = baseEndpoint + '/aggregates';
+export const transactionAggregatesApi = baseEndpoint + '/aggregates';
 
-export const transactionFeeAggregates = baseEndpoint + '/txfeeAggregates';
+export const transactionFeeAggregatesApi = baseEndpoint + '/txfeeAggregates';
